Fixed disagg_mag_dist_eps with mutex sources
