+++ openquake/hazardlib/source_reader.py
@@ -28,7 +28,7 @@ import numpy
 
 from openquake.baselib import parallel, general, hdf5, python3compat
 from openquake.hazardlib import nrml, sourceconverter, InvalidFile
-from openquake.hazardlib.valid import basename
+from openquake.hazardlib.valid import basename, fragmentno
 from openquake.hazardlib.lt import apply_uncertainties
 from openquake.hazardlib.geo.surface.kite_fault import kite_to_geom
 
@@ -75,16 +75,6 @@ def gzpik(obj):
     """
     gz = gzip.compress(pickle.dumps(obj, pickle.HIGHEST_PROTOCOL))
     return numpy.frombuffer(gz, numpy.uint8)
-
-
-def fragmentno(src):
-    "Postfix after :.; as an integer"
-    # in disagg/case-12 one has source IDs like 'SL_kerton:665!b16'
-    fragments = re.split('[:.;]', src.source_id)
-    if len(fragments) == 1:  # no fragment number, like in AELO for NZL
-        return -1
-    fragment = fragments[1].split('!')[0]  # strip !b16
-    return int(fragment)
 
 
 def mutex_by_grp(src_groups):
