+++ openquake/hazardlib/valid.py
@@ -1417,3 +1417,14 @@ def corename(src):
     """
     src = src if isinstance(src, str) else src.source_id
     return re.split('[!:;.]', src)[0]
+
+
+def fragmentno(src):
+    "Postfix after :.; as an integer"
+    # in disagg/case-12 one has source IDs like 'SL_kerton:665!b16'
+    fragments = re.split('[:.;]', src.source_id)
+    if len(fragments) == 1:  # no fragment number, like in AELO for NZL
+        return -1
+    fragment = fragments[1].split('!')[0]  # strip !b16
+    return int(fragment)
+
